apps/backend-core/requirements/qa.txt
# QA Environment - Pinned versions for reproducible builds
# Generated from base.txt + production dependencies

argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.1
attrs==25.4.0
boto3==1.42.43
botocore==1.42.43
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
crispy-bootstrap5==2025.6
cryptography==46.0.4
cssselect2==0.8.0
Deprecated==1.3.1
diff-match-patch==20241021
dj-rest-auth==7.0.2
Django==5.2.10
django-allauth==65.14.0
django-anymail==14.0
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-environ==0.12.0
django-extensions==4.1
django-filter==25.2
django-import-export==4.4.0
django-model-utils==5.0.0
django-multiselectfield==1.0.1
django-picklefield==3.4.0
django-q2==1.9.0
django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-spectacular==0.29.0
emoji==2.15.0
et_xmlfile==2.0.0
exa_py==2.3.0
Faker==40.1.2
fonttools==4.61.1
gevent==25.9.1
google-auth==2.48.0
google-auth-oauthlib==1.2.4
greenlet==3.3.0
gunicorn==23.0.0
hiredis==3.3.0
html5lib==1.1
idna==3.11
inflection==0.5.1
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kafka-python-ng==2.2.3
Markdown==3.10.1
msal==1.34.0
mux-python==5.1.2
numpy==2.4.2
oauthlib==3.3.1
openpyxl==3.1.5
packaging==26.0
pandas==2.2.3
pillow==12.1.0
psutil==7.2.2
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydyf==0.12.1
PyJWT==2.11.0
pymongo==4.16.0
pyotp==2.9.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-http-client==3.3.7
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
razorpay==2.0.0
redis==7.1.0
cent==5.2.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
sendgrid==6.12.5
six==1.17.0
sqlparse==0.5.5
starkbank-ecdsa==2.2.0
tablib==3.9.0
text-unidecode==1.3
tinycss2==1.5.1
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
weasyprint==68.1
webencodings==0.5.1
wrapt==2.1.1
zope.event==6.1
zope.interface==8.2
zopfli==0.4.0
